models/phi3.5-mini-instruct/requirements.txt
einops
accelerate
bitsandbytes
sentencepiece
torch
transformers
simple_ai_server>=0.2.1
#git+https://github.com/Youho99/simpleAI